fix(ui): pick CJK-safe terminal font fallbacks and bundle mono emoji
