+++ packages/ui/src/shell/terminal/assets.d.ts
@@ -2,3 +2,8 @@ declare module '*.woff2?inline' {
   const dataUri: string;
   export default dataUri;
 }
+
+declare module '*.ttf?inline' {
+  const dataUri: string;
+  export default dataUri;
+}
