+++ packages/ui/src/shell/terminal/live-terminal.tsx
@@ -1,64 +1,12 @@
-import ibmPlexMonoWoff2 from '@fontsource/ibm-plex-mono/files/ibm-plex-mono-latin-400-normal.woff2?inline';
 import { useEffect as useAbortableEffect } from 'foxact/use-abortable-effect';
 import { useLayoutEffect } from 'foxact/use-isomorphic-layout-effect';
 import { useRef } from 'react';
 import { createRestty } from 'restty';
 import type { PtyTransport } from 'restty/internal';
 import { cn } from '../../lib/cn';
+import { TERMINAL_FONT_SOURCES } from './fonts';
 import type { TerminalSession } from './session';
 import { applyTerminalTheme } from './terminal-theme';
-
-// restty renders on a GPU canvas with its own text shaper and needs raw font bytes; its default
-// `fontPreset: 'default-cdn'` fetches from cdn.jsdelivr.net, which the renderer CSP blocks. Bundle
-// IBM Plex Mono inline instead (no network, no CDN).
-const MONO_FONT = decodeDataUri(ibmPlexMonoWoff2);
-const TERMINAL_FONT_SOURCES: NonNullable<Parameters<typeof createRestty>[0]['fontSources']> = [
-  { type: 'buffer', data: MONO_FONT, label: 'IBM Plex Mono' },
-  // Fall back to the user's installed fonts for glyphs IBM Plex Mono lacks — Nerd/powerline
-  // icons, CJK, emoji — via the Local Font Access API when the host allows it. Match both full
-  // Nerd Font names and common abbreviated NF family names such as MesloLGS NF / CaskaydiaCove NF.
-  {
-    type: 'local',
-    matchers: [
-      'symbols nerd font',
-      'symbols nerd font mono',
-      'jetbrainsmono nerd font',
-      'jetbrains mono nerd font',
-      'fira code nerd font',
-      'hack nerd font',
-      'meslo lgm nerd font',
-      'meslo lgs nf',
-      'meslolgs nf',
-      'caskaydia',
-      'cascadia code nf',
-      'monaspace nerd font',
-      'nerd font mono',
-      'nerd font',
-      'powerline',
-    ],
-    label: 'symbols',
-  },
-  {
-    type: 'local',
-    matchers: ['pingfang', 'hiragino sans', 'microsoft yahei', 'noto sans cjk', 'source han sans'],
-    label: 'cjk',
-  },
-  {
-    type: 'local',
-    matchers: ['apple color emoji', 'segoe ui emoji', 'noto color emoji'],
-    label: 'emoji',
-  },
-  {
-    type: 'local',
-    matchers: ['sf mono', 'menlo', 'monaco', 'consolas', 'dejavu sans mono'],
-    label: 'mono',
-  },
-];
-
-function decodeDataUri(uri: string): ArrayBuffer {
-  const binary = atob(uri.slice(uri.indexOf(',') + 1));
-  return Uint8Array.from(binary, (ch) => ch.codePointAt(0) ?? 0).buffer;
-}
 
 // Every mounted LiveTerminal needs to re-theme on the same `.dark` class flip, but terminal tabs
 // stay mounted (visibility-toggled, not unmounted) while N tabs are open — so share a single
