+++ packages/workbench/src/assets.d.ts
@@ -2,3 +2,8 @@ declare module '*.woff2?inline' {
   const content: string;
   export default content;
 }
+
+declare module '*.ttf?inline' {
+  const content: string;
+  export default content;
+}
